Runtime/RTAcpi*: Some additions to create GpioInt() resource descriptors, bugref:10732 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108559 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/acpi/acpi.cpp
@@ -2447,7 +2447,7 @@ RTDECL(int) RTAcpiResourceAddGpioInt(RTACPIRES hAcpiRes, RTACPIRESGPIOMOD enmMod
     pb[14] = 23;                                                              /* Pin table offset[7:0]  */
     pb[15] = 0;                                                               /* Pin table offset[15:8] */
     pb[16] = 0;                                                               /* Resource Source Index  */
-    pb[17] = 23 + cPins * sizeof(*pau16Pins);                                 /* Resource Source Name Offset[7:0]  */
+    pb[17] = (uint8_t)(23 + cPins * sizeof(*pau16Pins));                      /* Resource Source Name Offset[7:0]  */
     pb[18] = 0;                                                               /* Resource Source Name Offset[15:8] */
     pb[19] = 0;                                                               /* Vendor Data Offset[7:0]           */
     pb[20] = 0;                                                               /* Vendor Data Offset[15:8]          */
